avoid constructing `Vector{Tuple{Type{...},Type{...}}}`
This commit adapts to the incoming upstream change of the Julia compiler.
More specifically this change is necessary after JuliaLang/julia#44725,
which fixes an internal bug of the compiler, which in turn makes the test
suite of SquidGame fail without this change (if you're interested, please
see <https://github.com/JuliaLang/julia/pull/44725#issuecomment-1078641874>).